Restore signature, make info colorized
