+++ src/client/providers/completionProvider.ts
@@ -1,5 +1,3 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
 'use strict';
 
 import * as vscode from 'vscode';
@@ -29,7 +27,15 @@ export class PythonCompletionItemProvider implements vscode.CompletionItemProvid
     }
 
     public async resolveCompletionItem(item: vscode.CompletionItem, token: vscode.CancellationToken): Promise<vscode.CompletionItem> {
-        item.documentation = await this.completionSource.getDocumentation(item, token);
+        if (!item.documentation) {
+            const itemInfos = await this.completionSource.getDocumentation(item, token);
+            if (itemInfos && itemInfos.length > 0) {
+                // Only filling documentation since tooltip already contains
+                // all the necessary information and is colorized via markdown.
+                item.documentation = itemInfos[0].tooltip;
+                item.detail = '';
+            }
+        }    
         return item;
     }
 }
